Refactoring: - use case calls makes calls for repository to check plane is in airport, instead of in the adapter - moving stubs into own package

+++ application/src/main/java/com/hanfak/airport/usecase/TakeOffUseCase.java
@@ -1,5 +1,6 @@
 package com.hanfak.airport.usecase;
 
+import com.hanfak.airport.domain.AirportStatus;
 import com.hanfak.airport.domain.plane.Plane;
 import com.hanfak.airport.domain.planetakeoffstatus.FailedPlaneTakeOffStatus;
 import com.hanfak.airport.domain.planetakeoffstatus.PlaneTakeOffStatus;
@@ -7,11 +8,13 @@ import com.hanfak.airport.domain.planetakeoffstatus.SuccessfulPlaneTakeOffStatus
 import com.hanfak.airport.domain.planetakeoffstatus.TakeOffFailureReason;
 import org.slf4j.Logger;
 
+import static com.hanfak.airport.domain.AirportStatus.IN_AIRPORT;
 import static com.hanfak.airport.domain.AirportStatus.NOT_IN_AIRPORT;
 import static com.hanfak.airport.domain.plane.PlaneStatus.FLYING;
 import static com.hanfak.airport.domain.planetakeoffstatus.FailedPlaneTakeOffStatus.failedPlaneTakeOffStatus;
 import static com.hanfak.airport.domain.planetakeoffstatus.PlaneTakeOffStatus.createPlaneTakeOffStatus;
 import static com.hanfak.airport.domain.planetakeoffstatus.SuccessfulPlaneTakeOffStatus.successfulPlaneTakeOffStatus;
+import static com.hanfak.airport.domain.planetakeoffstatus.TakeOffFailureReason.PLANE_COULD_NOT_TAKE_OFF;
 import static com.hanfak.airport.domain.planetakeoffstatus.TakeOffFailureReason.PLANE_IS_FLYING;
 import static com.hanfak.airport.domain.planetakeoffstatus.TakeOffFailureReason.PLANE_IS_NOT_AT_THE_AIRPORT;
 import static java.lang.String.format;
@@ -31,22 +34,27 @@ public class TakeOffUseCase {
     // Or Throw exception for failure case?? but information will get lost
     // Or Use a map
     // Or Can think about using railway programming??
-    public PlaneTakeOffStatus instructPlaneToTakeOff(Plane plane) { // TODO use planeID instead
+    public PlaneTakeOffStatus instructPlaneToTakeOff(Plane plane) {
         if (FLYING.equals(plane.planeStatus)) {
-            logger.info(format("Plane, '%s', cannot take off, status is '%s'", plane.planeId, plane.planeStatus.name()));
+            logger.info(format("Plane, '%s', cannot take off, status is '%s', it is not at the airport", plane.planeId, plane.planeStatus.name()));
             return createPlaneTakeOffStatus(null,
-                    getFailedPlaneTakeOffStatus(plane, PLANE_IS_FLYING));
+                    getFailedPlaneTakeOffStatus(plane, PLANE_IS_FLYING, NOT_IN_AIRPORT));
         }
 
-        try {
-            planeInventoryService.removePlane(plane);
+        // TODO try catch on this
+        if (planeInventoryService.planeIsPresentInAirport(plane)) {
+            try {
+                planeInventoryService.removePlane(plane);
+            } catch (IllegalStateException e) {
+                logger.error(format("Something went wrong removing the Plane, '%s', at the airport", plane.planeId), e);
+                return createPlaneTakeOffStatus(null, getFailedPlaneTakeOffStatus(plane, PLANE_COULD_NOT_TAKE_OFF, IN_AIRPORT));
+            }
             Plane flyingPlane = plane.fly();
             logger.info(format("Plane, '%s', has successfully left the airport", plane.planeId));
             return createPlaneTakeOffStatus(getSuccessfulPlaneTakeOffStatus(flyingPlane), null);
-        } catch (Exception e) {
-            logger.error(format("Plane, '%s', is not at airport", plane.planeId), e);
-            return createPlaneTakeOffStatus(null,
-                    getFailedPlaneTakeOffStatus(plane, PLANE_IS_NOT_AT_THE_AIRPORT));
+        } else {
+            logger.info(format("Plane, '%s', cannot take off, it is not at the airport", plane.planeId));
+            return createPlaneTakeOffStatus(null, getFailedPlaneTakeOffStatus(plane, PLANE_IS_NOT_AT_THE_AIRPORT, NOT_IN_AIRPORT));
         }
     }
 
@@ -56,7 +64,7 @@ public class TakeOffUseCase {
     }
 
     // TODO better name
-    private FailedPlaneTakeOffStatus getFailedPlaneTakeOffStatus(Plane plane, TakeOffFailureReason reason) {
-        return failedPlaneTakeOffStatus(plane.planeId, plane.planeStatus, NOT_IN_AIRPORT, reason);
+    private FailedPlaneTakeOffStatus getFailedPlaneTakeOffStatus(Plane plane, TakeOffFailureReason reason, AirportStatus airportStatus) {
+        return failedPlaneTakeOffStatus(plane.planeId, plane.planeStatus, airportStatus, reason);
     }
 }
